Course API tests partly works
