+++ src/main/resources/application-test.properties
@@ -2,7 +2,6 @@ springdoc.api-docs.enabled = true
 
 server.host.backend=${AU_BACKEND_HOST}
 server.host.frontend=${AU_FRONTEND_HOST}
-server.port=8900
 
 # GitHub API Key
 github.oauth.clientId = 60ccbeca3ab46b3d1699
