--- src/test/java/com/uu/au/CASTests.java
@@ -1,13 +1,0 @@
-package com.uu.au;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CASTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
